+++ crossbeam-utils/build.rs
@@ -19,7 +19,6 @@ include!("build-common.rs");
 
 fn main() {
     println!("cargo:rerun-if-changed=no_atomic.rs");
-    println!("cargo:rustc-check-cfg=cfg(crossbeam_no_atomic,crossbeam_sanitize_thread)");
 
     let target = match env::var("TARGET") {
         Ok(target) => convert_custom_linux_target(target),