+++ rmp-serde/src/lib.rs
@@ -28,9 +28,14 @@
 //! # Examples
 //!
 //! ```
+//! extern crate serde;
+//! #[macro_use]
+//! extern crate serde_derive;
+//! extern crate rmp_serde as rmps;
+//!
 //! use std::collections::HashMap;
 //! use serde::{Deserialize, Serialize};
-//! use rmp_serde::{Deserializer, Serializer};
+//! use rmps::{Deserializer, Serializer};
 //!
 //! #[derive(Debug, PartialEq, Deserialize, Serialize)]
 //! struct Human {
@@ -53,15 +58,18 @@
 #![forbid(unsafe_code)]
 #![warn(missing_debug_implementations, missing_docs)]
 
+#[macro_use]
+extern crate serde;
+
 use std::fmt::{self, Display, Formatter};
 use std::str::{self, Utf8Error};
 
 use serde::de;
 use serde::{Deserialize, Serialize};
 
+pub use crate::decode::{from_read, Deserializer};
 #[allow(deprecated)]
 pub use crate::decode::from_read_ref;
-pub use crate::decode::{from_read, Deserializer};
 pub use crate::encode::{to_vec, to_vec_named, Serializer};
 
 pub use crate::decode::from_slice;
@@ -70,15 +78,9 @@ pub mod config;
 pub mod decode;
 pub mod encode;
 
-/// Hack used to serialize MessagePack Extension types.
-///
-/// A special `ExtStruct` type is used to represent
-/// extension types. This struct is renamed in serde.
-///
 /// Name of Serde newtype struct to Represent Msgpack's Ext
-/// Msgpack Ext: `Ext(tag, binary)`
-/// Serde data model: `_ExtStruct((tag, binary))`
-///
+/// Msgpack Ext: Ext(tag, binary)
+/// Serde data model: _ExtStruct((tag, binary))
 /// Example Serde impl for custom type:
 ///
 /// ```ignore
@@ -104,14 +106,12 @@ pub struct Raw {
 impl Raw {
     /// Constructs a new `Raw` from the UTF-8 string.
     #[inline]
-    #[must_use]
     pub fn new(v: String) -> Self {
         Self { s: Ok(v) }
     }
 
     /// DO NOT USE. See <https://github.com/3Hren/msgpack-rust/issues/305>
     #[deprecated(note = "This feature has been removed")]
-    #[must_use]
     pub fn from_utf8(v: Vec<u8>) -> Self {
         match String::from_utf8(v) {
             Ok(v) => Raw::new(v),
@@ -126,21 +126,18 @@ impl Raw {
 
     /// Returns `true` if the raw is valid UTF-8.
     #[inline]
-    #[must_use]
     pub fn is_str(&self) -> bool {
         self.s.is_ok()
     }
 
     /// Returns `true` if the raw contains invalid UTF-8 sequence.
     #[inline]
-    #[must_use]
     pub fn is_err(&self) -> bool {
         self.s.is_err()
     }
 
     /// Returns the string reference if the raw is valid UTF-8, or else `None`.
     #[inline]
-    #[must_use]
     pub fn as_str(&self) -> Option<&str> {
         match self.s {
             Ok(ref s) => Some(s.as_str()),
@@ -151,7 +148,6 @@ impl Raw {
     /// Returns the underlying `Utf8Error` if the raw contains invalid UTF-8 sequence, or
     /// else `None`.
     #[inline]
-    #[must_use]
     pub fn as_err(&self) -> Option<&Utf8Error> {
         match self.s {
             Ok(..) => None,
@@ -161,7 +157,6 @@ impl Raw {
 
     /// Returns a byte slice of this raw's contents.
     #[inline]
-    #[must_use]
     pub fn as_bytes(&self) -> &[u8] {
         match self.s {
             Ok(ref s) => s.as_bytes(),
@@ -171,14 +166,12 @@ impl Raw {
 
     /// Consumes this object, yielding the string if the raw is valid UTF-8, or else `None`.
     #[inline]
-    #[must_use]
     pub fn into_str(self) -> Option<String> {
         self.s.ok()
     }
 
     /// Converts a `Raw` into a byte vector.
     #[inline]
-    #[must_use]
     pub fn into_bytes(self) -> Vec<u8> {
         match self.s {
             Ok(s) => s.into_bytes(),
@@ -190,7 +183,7 @@ impl Raw {
 impl Serialize for Raw {
     fn serialize<S>(&self, se: S) -> Result<S::Ok, S::Error>
     where
-        S: serde::Serializer,
+        S: serde::Serializer
     {
         match self.s {
             Ok(ref s) => se.serialize_str(s),
@@ -271,13 +264,11 @@ pub struct RawRef<'a> {
 impl<'a> RawRef<'a> {
     /// Constructs a new `RawRef` from the UTF-8 string.
     #[inline]
-    #[must_use]
     pub fn new(v: &'a str) -> Self {
         Self { s: Ok(v) }
     }
 
     #[deprecated(note = "This feature has been removed")]
-    #[must_use]
     pub fn from_utf8(v: &'a [u8]) -> Self {
         match str::from_utf8(v) {
             Ok(v) => RawRef::new(v),
@@ -291,21 +282,18 @@ impl<'a> RawRef<'a> {
 
     /// Returns `true` if the raw is valid UTF-8.
     #[inline]
-    #[must_use]
     pub fn is_str(&self) -> bool {
         self.s.is_ok()
     }
 
     /// Returns `true` if the raw contains invalid UTF-8 sequence.
     #[inline]
-    #[must_use]
     pub fn is_err(&self) -> bool {
         self.s.is_err()
     }
 
     /// Returns the string reference if the raw is valid UTF-8, or else `None`.
     #[inline]
-    #[must_use]
     pub fn as_str(&self) -> Option<&str> {
         match self.s {
             Ok(s) => Some(s),
@@ -316,7 +304,6 @@ impl<'a> RawRef<'a> {
     /// Returns the underlying `Utf8Error` if the raw contains invalid UTF-8 sequence, or
     /// else `None`.
     #[inline]
-    #[must_use]
     pub fn as_err(&self) -> Option<&Utf8Error> {
         match self.s {
             Ok(..) => None,
@@ -326,7 +313,6 @@ impl<'a> RawRef<'a> {
 
     /// Returns a byte slice of this raw's contents.
     #[inline]
-    #[must_use]
     pub fn as_bytes(&self) -> &[u8] {
         match self.s {
             Ok(s) => s.as_bytes(),