+++ strsim/tests/lib.rs
@@ -1,30 +1,13 @@
 extern crate strsim;
 
-use strsim::{
-    damerau_levenshtein, hamming, jaro, jaro_winkler, levenshtein, normalized_damerau_levenshtein,
-    normalized_levenshtein, osa_distance,
-};
-
-macro_rules! assert_delta {
-    ($x:expr, $y:expr) => {
-        assert_delta!($x, $y, 1e-5);
-    };
-    ($x:expr, $y:expr, $d:expr) => {
-        if ($x - $y).abs() > $d {
-            panic!(
-                "assertion failed: actual: `{}`, expected: `{}`: \
-                actual not within < {} of expected",
-                $x, $y, $d
-            );
-        }
-    };
-}
+use strsim::{hamming, levenshtein, normalized_levenshtein, osa_distance,damerau_levenshtein,
+             normalized_damerau_levenshtein, jaro, jaro_winkler};
 
 #[test]
 fn hamming_works() {
     match hamming("hamming", "hammers") {
         Ok(distance) => assert_eq!(3, distance),
-        Err(why) => panic!("{:?}", why),
+        Err(why) => panic!("{:?}", why)
     }
 }
 
@@ -35,7 +18,7 @@ fn levenshtein_works() {
 
 #[test]
 fn normalized_levenshtein_works() {
-    assert_delta!(0.57142, normalized_levenshtein("kitten", "sitting"));
+    assert!((normalized_levenshtein("kitten", "sitting") - 0.57142).abs() < 0.00001);
 }
 
 #[test]
@@ -50,22 +33,17 @@ fn damerau_levenshtein_works() {
 
 #[test]
 fn normalized_damerau_levenshtein_works() {
-    assert_delta!(
-        0.27272,
-        normalized_damerau_levenshtein("levenshtein", "löwenbräu")
-    );
+    assert!((normalized_damerau_levenshtein("levenshtein", "löwenbräu") - 0.27272).abs() < 0.00001);
 }
 
 #[test]
 fn jaro_works() {
-    assert_delta!(
-        0.392,
-        jaro("Friedrich Nietzsche", "Jean-Paul Sartre"),
-        0.001
-    );
+    assert!((0.392 - jaro("Friedrich Nietzsche", "Jean-Paul Sartre")).abs() <
+            0.001);
 }
 
 #[test]
 fn jaro_winkler_works() {
-    assert_delta!(0.866, jaro_winkler("cheeseburger", "cheese fries"), 0.001);
+    assert!((0.911 - jaro_winkler("cheeseburger", "cheese fries")).abs() <
+            0.001);
 }