+++ zeroize/tests/zeroize_derive.rs
@@ -1,5 +1,4 @@
 //! Integration tests for `zeroize_derive` proc macros
-
 #![cfg(feature = "zeroize_derive")]
 
 use zeroize::{Zeroize, ZeroizeOnDrop};
@@ -319,7 +318,6 @@ fn derive_deref() {
 
 #[test]
 #[cfg(feature = "alloc")]
-#[allow(dead_code)]
 fn derive_zeroize_on_drop_generic() {
     #[derive(ZeroizeOnDrop)]
     struct Y<T: Zeroize>(Box<T>);
@@ -329,7 +327,6 @@ fn derive_zeroize_on_drop_generic() {
 }
 
 #[test]
-#[allow(dead_code)]
 fn derive_zeroize_unused_param() {
     #[derive(Zeroize)]
     struct Z<T> {
@@ -340,7 +337,6 @@ fn derive_zeroize_unused_param() {
 }
 
 #[test]
-#[allow(dead_code)]
 // Issue #878
 fn derive_zeroize_with_marker() {
     #[derive(ZeroizeOnDrop, Zeroize)]
@@ -349,7 +345,6 @@ fn derive_zeroize_with_marker() {
         field: Option<A>,
     }
 
-    #[allow(dead_code)]
     trait Secret: ZeroizeOnDrop + Zeroize {}
 
     impl<A: Marker> Secret for Test<A> {}
@@ -358,7 +353,6 @@ fn derive_zeroize_with_marker() {
 }
 
 #[test]
-#[allow(dead_code)]
 // Issue #878
 fn derive_zeroize_used_param() {
     #[derive(Zeroize)]